requirements.txt
boto3==1.26.29
botocore==1.29.29
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
comtypes==1.1.14
decorator==5.1.1
dialogflow-fulfillment==0.4.4
Django==2.2.4
djangorestframework
dragonfly2==0.35.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
jmespath==1.0.1
json-rpc==1.13.0
lark-parser==0.8.9
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==22.0
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
setuptools==63.2.0
six==1.16.0
sqlparse==0.4.3
twilio==7.15.4
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.13
webhook-client==1.0.2
Werkzeug==2.2.2
ssml-builder==1.1.1
voximplant-apiclient==1.3.7
openai
django-crispy-forms
sendgrid
pydub
ffmpeg-python
ffmpeg
google-cloud-dialogflow-cx
pymongo
dnspython
mongoengine
nltk
langchain
chromadb
PyPDF2